add course finishing message

+++ app/models/user.py
@@ -1,7 +1,6 @@
 from datetime import datetime
-from enum import IntEnum
 
-from sqlalchemy import Column, Integer, String, Date, DateTime, ForeignKey, Boolean, Text, Time
+from sqlalchemy import Column, Integer, String, Date, DateTime, ForeignKey, Boolean, Text, Time, BigInteger
 from sqlalchemy.orm import relationship
 
 from app.config import app_config
@@ -12,7 +11,7 @@ class User(Base):
     __tablename__ = "user"
 
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, nullable=False, index=True)
+    telegram_id = Column(BigInteger, unique=True, nullable=False, index=True)
     phone = Column(String(20), unique=True, nullable=True, index=True)
     first_name = Column(String(100), nullable=True)
     last_name = Column(String(100), nullable=True)
